Correct Task 0

+++ 0x12-singly_linked_lists/main.c
@@ -13,7 +13,7 @@ int main(void)
 	list_t *head;
 	list_t *new;
 	list_t hello = {"World", 5, NULL};
-	long unsigned int n;
+	size_t n;
 
 	head = &hello;
 	new = malloc(sizeof(list_t));
